Created directories for pages, moved shared components to /components
