+++ src/client/components/AuthedNavbar.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { NavLink, Outlet } from "react-router-dom";
+import { NavLink } from "react-router-dom";
 
 const AuthedNavbar = () => {
   return (
