+++ src/client/components/ProtectedRoute.jsx
