+++ src/client/components/SharedLayout.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Link, Outlet } from "react-router-dom";
+import { Outlet } from "react-router-dom";
 import Navbar from "./Navbar.jsx";
 
 const SharedLayout = () => {
