--- src/client/containers/MainContainer.jsx
@@ -1,16 +1,0 @@
-import React from "react";
-
-const MainContainer = () => {
-  // add pertinent state here
-
-  return (
-    <div className="container">
-      <div className="outerBox">
-        <h1 id="header">MegaMarket Loyalty Cards</h1>
-        {/* Start adding components here... */}
-      </div>
-    </div>
-  );
-};
-
-export default MainContainer;
