+++ src/client/pages/Home.jsx
@@ -1,6 +1,4 @@
 import React from "react";
-import { Link, Outlet } from "react-router-dom";
-import Navbar from "./Navbar.jsx";
 
 const Home = () => {
   return (
